CERN - Add NetClusterFreeIps scheduler filter
This scheduler filter evaluates if there are available
IP addresses in the host netcluster to start a VM.
If not the host is excluded.

In order to prevent parallel IP allocation a configuration
option is available "netcluster_min_free_ips" in order to define a
minimum number of available IP addresses in the netcluster.

JIRA: OS-97