live-migration use policy
Because the DB function requires an admin context, if you set a policy for
the live-migration function (compute_extension:admin_actions:migrateLive)
to not require admin the user must still be an admin.

If you set the policy to not require admin for live-migration then non-admin
users should be able to run live-migration.

This change makes it so that the administrator can set the policy
to non-admin if that's appropriate for their setup

Reference: https://review.openstack.org/26972/